+++ views/common/copy.php
@@ -6,12 +6,13 @@
  * @license https://www.humhub.com/licences
  */
 
-use humhub\libs\Html;
 use humhub\modules\custom_pages\models\CustomPage;
 use humhub\modules\ui\form\widgets\ActiveForm;
+use humhub\modules\ui\view\components\View;
 use humhub\widgets\ModalButton;
 use humhub\widgets\ModalDialog;
 
+/* @var $this View */
 /* @var $page CustomPage */
 ?>
 <?php ModalDialog::begin([
@@ -31,17 +32,12 @@ use humhub\widgets\ModalDialog;
         </div>
     <?php endif; ?>
 
-    <?= $form->field($page, 'visibility')->radioList($page->getVisibilitySelection()) ?>
+    <?= $this->render('edit_visibility', ['page' => $page, 'form' => $form]) ?>
     <?= $form->field($page, 'target')->dropDownList($page->getAvailableTargetOptions()) ?>
 </div>
 <div class="modal-footer">
     <?= ModalButton::cancel() ?>
     <?= ModalButton::submitModal(null, Yii::t('CustomPagesModule.view', 'Copy')) ?>
 </div>
-<script <?= Html::nonce() ?>>
-    $('input[type="radio"][name="CustomPage[visibility]"]').click(function () {
-        $('.infoAdminOnly').toggle($(this).val() == <?= CustomPage::VISIBILITY_ADMIN ?>);
-    });
-</script>
 <?php ActiveForm::end() ?>
 <?php ModalDialog::end() ?>